requirements.txt
# Web UI + RAG API (see also pubmed_ingest.py for ingestion)
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic-settings>=2.0
langchain-core>=0.3
langchain-openai>=0.2
langchain-pinecone>=0.2
pinecone>=5.0
requests>=2.28
cohere>=5.0
apscheduler>=3.10
aiosqlite>=0.19
